LayoutTests/imported/w3c/web-platform-tests/resource-timing/resource_TAO_multi-expected.txt

Blocked access to external URL http://www.localhost:8800/resource-timing/resources/TAOResponse.py?tao=multi
CONSOLE MESSAGE: line 13: NetworkError:  A network error occurred.
Description

This test validates that for a cross origin resource, the timing allow check algorithm will pass when the HTTP response has multiple Timing-Allow-Origin header fields and the subsequent field value is separated by a comma.


PASS window.performance is defined 
PASS window.performance.getEntriesByType is defined 
